code/WebAPI/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.23.2
sqlalchemy
pymysql==1.1.1
pydantic[email]
python-dotenv==1.0.0
bleach